added filter-repetitive to settings
